Initial Commit
Hide a limited number of specified posts from the homepage

We keep a list of post ID's that is limited in size to keep
the database queries reasonable. To remove the posts, we
use a NOT IN query with the array of post ID's. When the array
gets to long, we start pushing out the oldest items.